Restructured LinearRegression class to be more safe - Created a protocol LinearRegression - Implemented two separate conforming types: class GradientDescentLinearRegression and class SingleValueDecompositionLinearRegression. - Updated LinearRegression notebook
